[ALPS04239425] Migration: Fix kernel build error
[Detail]
1.Merge power part.
2.Using correct mt_secure_call function.
3.Temp disable SMI.
4.Replace pm_wakeup with wake_lock.
5.Merge mmc part.
6.Add thermal_zones part.
7.Fix ccu_trigger_i2c part.
8.Fix diplay part.
9.Merge gpu part.
10.Fix pmic part.

MTK-Commit-Id: 9e6549e431d6152f8b41982ccb1c12827840918f

CR-Id: ALPS04239425
Feature: [Module]Kernel Maintenance

+++ drivers/input/keyboard/mediatek/mt6763/hal_kpd.c
@@ -303,7 +303,7 @@ void kpd_wakeup_src_setting(int enable)
 }
 
 /********************************************************************/
-void kpd_init_keymap(u16 keymap[])
+void kpd_init_keymap(u32 keymap[])
 {
 	int i = 0;
